Automatic numbering
There are still some issues with the statutes' hierarchy. For instance, paragraphs
are set on a 4th level, but sometimes they actually are on a 3rd level
(so they are displayed like this: 1.2.0.1, 1.2.0.2, ... instead of
1.2.1, 1.2.2, ...).

When this will be fixed, we will have to remove the hard-coded numbering.

By the way, I put statutes locally, in a directory (assets/statutes/)
for testing purposes.